# stream-1-backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: FitForecast Stream 1 Backend API
  version: '0.1.0'
  description: Minimal OpenAPI spec for the FitForecast backend (entries, feelings, trends, insights)
servers:
  - url: http://localhost:3000
paths:
  /entries:
    post:
      summary: Create a new log entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id:
                  type: string
                  format: uuid
                type:
                  type: string
                  enum: [workout, meal]
                raw_text:
                  type: string
                occurred_at:
                  type: string
                  format: date-time
      responses:
        '201':
          description: Created
    get:
      summary: List entries
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
        - in: query
          name: type
          schema:
            type: string
            enum: [workout, meal]
      responses:
        '200':
          description: OK

  /entries/{id}/feelings:
    post:
      summary: Create a feeling snapshot for an entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                when:
                  type: string
                  enum: [pre, post]
                valence:
                  type: integer
                energy:
                  type: integer
                stress:
                  type: integer
      responses:
        '201':
          description: Created

  /trends:
    get:
      summary: Get baseline trends and recent entries
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
        - in: query
          name: window_days
          schema:
            type: integer
            enum: [7, 30, 365]
      responses:
        '200':
          description: OK

  /insights:
    get:
      summary: List active insights
      parameters:
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK

components: {}